% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{binormal}
\alias{binormal}
\title{eight_schools}
\format{An S4 object of class \code{stanfit}}
\source{
\url{mc-stan.org}
}
\description{
Output from Stan demo "binormal".
}
\seealso{
\code{\link[rstan]{stan_demo}}
}
\keyword{datasets}

