% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/deploy_shinystan.R
\name{deploy_shinystan}
\alias{deploy_shinystan}
\title{Deploy to shinyapps.io}
\usage{
deploy_shinystan(sso, account, appName, appDir = getwd(),
  ppcheck_data = NULL)
}
\arguments{
\item{sso}{The \code{shinystan} object for the model you want to use.}

\item{account}{ShinyApps account username. Not required if only one
ShinyApps account is configured on the system.
See \code{\link[shinyapps]{deployApp}} and \code{\link[shinyapps]{accounts}}.}

\item{appName}{The name to use for the application as a character string.}

\item{appDir}{Path to shinystan_for_shinyapps library. Defaults to working directory.
See \strong{Details}.}

\item{ppcheck_data}{Optional vector of observations to use for graphical posterior
predictive checking. Not used if missing or `NULL`.}
}
\description{
Requires a ShinyApps account. Visit http://www.shinyapps.io/ to sign up.
}
\details{
In order to deploy a shinyStan app to shinyapps.io you first
need to download the \code{shinystan_for_shinyapps} library, which is
available at https://github.com/stan-dev/shinystan/releases. You will
also need to set up a ShinyApps account (see http://www.shinyapps.io/).
}
\note{
With one exception, all shinyStan features should work properly on shinyapps.io.
The exception is the trivariate 3D scatterplot, which is not available in shinyStan apps
on shinyapps.io at this time.
}
\examples{
\dontrun{

# For this example assume my_sso is the name of the shinystan object
# you want to use and that your ShinyApps username is 'username'.

# if we first set the working directory to be 'shinystan_for_shinyapps'
# we don't need to specify the appDir argument
setwd(Path to shinystan_for_shinyapps folder)
deploy_shinystan(my_sso, appName = "my_shinystan_app", account = "username")
}
}
\seealso{
\code{\link[shinyapps]{deployApp}}, \code{\link[shinyapps]{accounts}}
}

