% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_parameters.R
\name{drop_parameters}
\alias{drop_parameters}
\title{Drop parameters from a shinystan object}
\usage{
drop_parameters(sso, pars)
}
\arguments{
\item{sso}{The shinystan object to use.}

\item{pars}{A character vector of parameter names. If the names of any 
non-scalar (e.g. vector, matrix) parameters are included in \code{pars} all
elements will be removed. It is not possible to remove only a subset of 
elements of non-scalar parameters.}
}
\value{
\code{sso} with \code{pars} dropped.
}
\description{
Drop parameters from a shinystan object
}

