% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eight_schools.R
\docType{data}
\name{eight_schools}
\alias{eight_schools}
\title{shinyStan demo: Eight Schools}
\format{An S4 object of class \code{shinystan}}
\usage{
eight_schools
}
\description{
A shinyStan object for the Stan "Eight Schools" demo model
}
\keyword{datasets}

