% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate a report with diagnostics for a shinystan object}
\usage{
generate_report(sso, n_param = 3, output_format = "html_document",
  view = TRUE)
}
\arguments{
\item{sso}{A \code{\link[=as.shinystan]{shinystan object}}.}

\item{n_param}{On how many parameters do you want the report to be?}

\item{output_format}{What type of report would you like? The options are
'html_document', 'pdf_document' and 'word_document'.}

\item{view}{Do you want to open the report after it is generated?}
}
\value{
A report is generated and the path where it is stored is printed.
}
\description{
The report displays diagnostics information about the worst parameters in 
your model, based on effective sample size.
}
\examples{
 \dontrun{
# Using example shinystan object 'eight_schools'
generate_report(eight_schools)
}

}
