% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/include_model_code.R
\name{include_model_code}
\alias{include_model_code}
\title{Add model code to a \code{shinystan} object}
\usage{
include_model_code(sso, code)
}
\arguments{
\item{sso}{A \code{shinystan} object.}

\item{code}{The code you want to add. See \strong{Details} below for
formatting instructions.}
}
\value{
The \code{shinystan} object \code{sso} with \code{code} in the
\code{model_code} slot.
}
\description{
Add model code to a \code{shinystan} object
}
\details{
\code{code} should be a character string that can be used as an argument
to \code{cat}. See \strong{Examples}, below.
}
\note{
This is intended for users who did not run their models using \pkg{rstan}.
For \pkg{rstan} users the model code will be automatically available.
}
\examples{
\dontrun{
# Some JAGS-style code we might want to add
my_code <- "
 model {
   for (i in 1:length(Y)) {
     Y[i] ~ dpois(lambda[i])
     log(lambda[i]) <- inprod(X[i,], theta[])
   }
   for (j in 1:J) {
     theta[j] ~ dt(0.0, 1.0, 1.0)
   }
 }
"

# Add the code to a shinystan object X
X <- include_model_code(X, my_code)
launch_shinystan(X) # code is visible in the Model Code tab

}
}
\seealso{
\code{cat}
}

