% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/launch_shinystan.R
\name{launch_shinystan}
\alias{launch_shinystan}
\title{Launch shinyStan app}
\usage{
launch_shinystan(object, ...)
}
\arguments{
\item{object}{An object of class \code{shinystan} or \code{stanfit}. See
\code{\link[shinyStan]{as.shinystan}} for converting to a
\code{shinystan} object (sso).}

\item{...}{Optional arguments to pass to \code{\link[shiny]{runApp}}. See
Details (RStudio users in particular).}
}
\value{
An S4 object of class \code{shinystan}.
}
\description{
Launch shinyStan app
}
\details{
Unless you are using RStudio, \code{launch_shinystan} will open the
  app in your system's default web browser. For RStudio users
  \strong{shinyStan} will launch in RStudio's (pop-up) Viewer pane by
  default. If you prefer to use \strong{shinyStan} in your web browser (or if
  you are having trouble with the RStudio Viewer pane) you can
  specify \code{launch.browser = TRUE} in \code{...}. Alternatively, if
  \strong{shinyStan} is open in the RStudio Viewer pane you can click on
  'Open in Browser' (at the top left of the Viewer pane).
}
\examples{
\dontrun{
# If X is a stanfit object (or shinystan object (sso))
X_sso <- launch_shinystan(X)

# If X is not an sso or stanfit object
X_sso <- launch_shinystan(as.shinystan(X, model_name = "Example"))
}
}
\seealso{
\code{\link[shinyStan]{as.shinystan}},
  \code{\link[shinyStan]{launch_shinystan_demo}}
}

