% Generated by roxygen2 (4.0.2): do not edit by hand
\name{launch_SHINYstan2}
\alias{launch_SHINYstan2}
\title{Launch SHINYstan app (experimental)}
\usage{
launch_SHINYstan2(object)
}
\arguments{
\item{object}{An object of class \code{stanfit} or \code{SHINYstanfit}.}
}
\value{
In addition to launching the app this function returns S4 object of
class \code{SHINYstanfit}.
}
\description{
This function will create an S4 object of class \code{SHINYstanfit} in the
Global Environment and launch the SHINYstan app.
}
\seealso{
\code{\link[rstan]{stanfit-class}}, \code{\link[SHINYstan]{SHINYstanfit-class}}
}

