% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mcmc2shinystan}
\alias{mcmc2shinystan}
\title{Convert an object of class \code{mcmc.list} to class \code{SHINYstanfit}}
\usage{
mcmc2shinystan(X, info = list(model_name = deparse(substitute(X)), burnin = 0,
  param_dims = list()))
}
\arguments{
\item{X}{An mcmc.list}

\item{info}{An optional named list of additional information. See \strong{Details}
below.}
}
\description{
Convert an object of class \code{mcmc.list} to class \code{SHINYstanfit}
}
\details{
\code{info} is an optional named list containing
\describe{
  \item{\code{model_name}}{A name for the model}
  \item{\code{burnin}}{The number of burnin iterations. Not needed if the
  samples don't include any of the burnin.}
  \item{\code{param_dims}}{The dimensions for all parameters. A named list.}
}
}
\seealso{
\code{\link[coda]{mcmc.list}}
}

