% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sso-helpers.R
\name{notes}
\alias{notes}
\title{Additional or replacement notes}
\usage{
notes(sso, note, replace = FALSE)
}
\arguments{
\item{sso}{The shinystan object to use.}

\item{note}{Optional. A character string containing a note to add to any
existing notes or replace existing notes, depending on the value of 
\code{replace}.}

\item{replace}{If \code{TRUE} the existing notes are overwritten by 
\code{note} if \code{note} is specified. If \code{FALSE} (the default) 
if \code{note} is specified then its content is appended to the existing
notes.}
}
\value{
If \code{note} is missing then any existing notes stored in 
  \code{sso} are returned as a character string. If \code{note} is specified
  then an updated shinystan object is returned with either \code{note} added
  to the previous notes (if \code{replace=FALSE}) or overwritten by
  \code{note} (if \code{replace = TRUE}).
}
\description{
Notes are viewable on ShinyStan's Notepad page
}
\examples{
# Using example shinystan object 'eight_schools'
sso <- eight_schools

# View existing notes
notes(sso)

# Add a note to the existing notes
sso <- notes(sso, "New note")
notes(sso)
cat(notes(sso))

# Replace existing notes
sso <- notes(sso, "replacement note", replace = TRUE)
notes(sso)
 
}
\seealso{
\code{\link{as.shinystan}}
}

