% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sso-metadata.R
\name{rename_model}
\alias{rename_model}
\title{rename_model (deprecated)}
\usage{
rename_model(sso, new_model_name)
}
\arguments{
\item{sso, new_model_name}{Use the \code{\link{model_name}} function instead.}
}
\description{
This function is deprecated and will be removed in a future release. Please 
use the \code{\link{model_name}} function instead.
}
\keyword{internal}

