% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{schools}
\alias{schools}
\title{schools}
\format{An S4 object of class \code{stanfit}}
\source{
\url{mc-stan.org}
}
\description{
Output from Stan demo "schools"
}
\seealso{
\code{\link[rstan]{stan_demo}}
}
\keyword{datasets}

