% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinystan-class.R
\docType{class}
\name{shinystan-class}
\alias{shinystan}
\alias{shinystan-class}
\title{S4 shinystan objects}
\description{
See \code{\link{as.shinystan}} for documentation on creating 
  shinystan objects and \code{\link{eight_schools}} for an example object.
}

