% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stan2shinystan}
\alias{stan2shinystan}
\title{Convert an object of class \code{stanfit} (\pkg{rstan}) to class \code{shinystan}}
\usage{
stan2shinystan(stanfit, notes, make = FALSE)
}
\arguments{
\item{stanfit}{A \code{stanfit} object.}

\item{notes}{A character string.}

\item{make}{.}
}
\description{
Convert an object of class \code{stanfit} (\pkg{rstan}) to class \code{shinystan}
}
\seealso{
\code{\link[rstan]{stan}}
}

