% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stanfit_to_SHINYstanfit}
\alias{stanfit_to_SHINYstanfit}
\title{Convert an object of class \code{stanfit} to class \code{SHINYstanfit}}
\usage{
stanfit_to_SHINYstanfit(stanfit, make = FALSE)
}
\description{
Convert an object of class \code{stanfit} to class \code{SHINYstanfit}
}

