% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sso.R
\name{update_sso}
\alias{update_sso}
\title{Update an object created by the previous version of shinystan}
\usage{
update_sso(sso)
}
\arguments{
\item{sso}{An old shinystan object to update.}
}
\value{
\code{sso}, updated.
}
\description{
If you encounter any errors when using an old shinystan object (\code{sso}) 
created by the previous version of \pkg{shinystan} you might need to run
\code{update_sso}. If \code{update_sso} does not resolve the problem and 
you still have the object (e.g. stanfit,  stanreg, mcmc.list) from which 
\code{sso} was originally created, you can create a new shinystan object 
using \code{\link{as.shinystan}}.
}

