% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.shinystan}
\alias{as.shinystan}
\alias{is.shinystan}
\title{\code{shinystan} Objects}
\usage{
as.shinystan(X, ...)

is.shinystan(Y)
}
\arguments{
\item{X}{A \code{stanfit}, \code{mcmc.list}, or 3D array of
posterior samples. If \code{X} is not a \code{stanfit} or \code{mcmclist} object
then the dimensions of the 3D array must correspond to iterations, chains,
and parameters, in that order.}

\item{...}{Additional arguments. See \strong{Details}, below, for instructions.}

\item{Y}{An object to test.}
}
\value{
For \code{as.shinystan} an object of class \code{shinystan}. For
\code{is.shinystan} a logical value indicating whether the tested object
is a \code{shinystan} object.
}
\description{
\code{shinystan} Objects
}
\details{
If \code{X} is a \code{stanfit} object then no additional arguments
should be specified in \code{...}. If \code{X} is an \code{mcmc.list} or
3D array of samples then the following arguments can be specified:
\describe{
  \item{\code{model_name}}{A character string giving a name for the model.}
  \item{\code{burnin}}{The number of warmup/burnin iterations.
  Not needed if the samples don't include any of the burnin.}
  \item{\code{param_dims}}{The dimensions for all parameters. A named list.}
}
}

