% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/deploy_shinystan.R
\name{deploy_shinystan}
\alias{deploy_shinystan}
\title{Deploy to shinyapps.io}
\usage{
deploy_shinystan(sso, directory, y, ...)
}
\arguments{
\item{sso}{The \code{shinystan} object to use.}

\item{directory}{Path to shinystan_for_shinyapps repository. See \strong{Details}.}

\item{y}{Optional. Vector of observations to use for graphical posterior
predictive checking.}

\item{...}{Arguments to pass to \code{\link[shinyapps]{deployApp}}.}
}
\description{
Deploy to shinyapps.io
}
\details{
In order to deploy a shinyStan app to shinyapps.io you first
need to clone the \code{shinystan_for_shinyapps} GitHub repository, which
contains a modified version of the shinyStan source code required for deploying
to shinyapps.io.
}
\seealso{
\code{\link[shinyapps]{deployApp}}
}

