% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_parameters.R
\name{drop_parameters}
\alias{drop_parameters}
\title{Drop parameters from a shinystan objects}
\usage{
drop_parameters(sso, pars)
}
\arguments{
\item{sso}{A shinystan object.}

\item{pars}{A character vector of parameter names. For a non-scalar parameter
\code{theta}, all elements will be removed if \code{"theta"} is included in
\code{pars}. A subset of the elements of \code{"theta"} can be removed by 
explicitly including them in \code{pars}, e.g. \code{pars <- c("theta[1]",
"theta[4]")}.}
}
\value{
\code{sso} with \code{pars} dropped.
}
\description{
Drop parameters from a shinystan objects
}

