% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eight_schools_demo.R
\docType{data}
\name{eight_schools}
\alias{eight_schools}
\title{ShinyStan demo: Eight Schools}
\format{An S4 object of class \code{shinystan}}
\usage{
eight_schools
}
\description{
A shinystan object for the Stan "Eight Schools" demo model
}
\seealso{
\code{\link{launch_shinystan_demo}}, \code{\link{as.shinystan}}
}
\keyword{datasets}

