% Generated by roxygen2 (4.0.2): do not edit by hand
\name{launch_shinystan}
\alias{launch_shinystan}
\title{Launch SHINYstan app}
\usage{
launch_shinystan(object)
}
\arguments{
\item{object}{An object of class \code{shinystan} or \code{stanfit}.}
}
\value{
If \code{object} is a \code{stanfit} object then, in addition to launching
the SHINYstan app, this function creates an object of class \code{shinystan} in
the Global Environment.
}
\description{
Launch SHINYstan app
}
\examples{
\dontrun{
# If X is a shinystan object or stanfit object then to launch the app
# just run
launch_shinstan(X)

# If X is an mcmc.list object or 3D array then to launch the app first
# convert X to a shinystan object using the as.shinystan function
X_shinystan <- as.shinystan(X)
launch_shinystan(X_shinystan)

}
}

