% Generated by roxygen2 (4.0.2): do not edit by hand
\name{launch_SHINYstan}
\alias{launch_SHINYstan}
\title{Launch SHINYstan app}
\usage{
launch_SHINYstan(stanfit)
}
\arguments{
\item{stanfit}{An S4 object of class \code{\link[rstan]{stanfit-class}} (\pkg{rstan}).}
}
\description{
This function will create an S4 object of class \code{SHINYstanfit} in the Global Environment
and then launch the SHINYstan app.
}
\seealso{
\code{\link[rstan]{stan}}
}

