% Generated by roxygen2 (4.0.2): do not edit by hand
\name{launch_SHINYstan}
\alias{launch_SHINYstan}
\title{Launch SHINYstan app}
\usage{
launch_SHINYstan(object)
}
\arguments{
\item{object}{An object of class \code{stanfit} or \code{SHINYstanfit}.}
}
\description{
This function will create an S4 object of class \code{SHINYstanfit} in the Global Environment
and then launch the SHINYstan app.
}
\seealso{
\code{\link[rstan]{stanfit-class}}, \code{\link[SHINYstan]{SHINYstanfit-class}}
}

