% Generated by roxygen2 (4.0.2): do not edit by hand
\name{launch_shinystan}
\alias{launch_shinystan}
\title{Launch SHINYstan app}
\usage{
launch_shinystan(object, ...)
}
\arguments{
\item{object}{An object of class \code{SHINYstanfit} or \code{stanfit}.}

\item{...}{Further arguments to be passed to \code{\link[SHINYstan]{stan2shinystan}}
if \code{object} is a \code{stanfit} object.}
}
\value{
If \code{object} is a \code{stanfit} object, in addition to launching
the app an object of class \code{SHINYstanfit} is returned.
}
\description{
This function will create an S4 object of class \code{SHINYstanfit} in the
Global Environment and launch the SHINYstan app.
}
\seealso{
\code{\link[rstan]{stanfit-class}}, \code{\link[SHINYstan]{SHINYstanfit-class}}
}

