% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/launch_shinystan.R
\name{launch_shinystan}
\alias{launch_shinystan}
\title{ShinyStan app}
\usage{
launch_shinystan(object, rstudio = FALSE, ...)
}
\arguments{
\item{object}{An object of class shinystan, stanfit, or
\code{stanreg}. See \code{\link{as.shinystan}} for converting other
objects to a shinystan object (sso).}

\item{rstudio}{For RStudio users, should the app launch in RStudio's Viewer?
The default (FALSE) is to launch the app in the default web browser.}

\item{...}{Optional arguments to pass to \code{\link[shiny]{runApp}}.}
}
\value{
An S4 shinystan object.
}
\description{
ShinyStan app
}
\examples{
\dontrun{
# If X is a stanfit object (or shinystan object (sso))
X_sso <- launch_shinystan(X)

# If X is not an sso or stanfit object
X_sso <- launch_shinystan(as.shinystan(X, model_name = "Example"))
}
}
\seealso{
\code{\link{as.shinystan}},
  \code{\link{launch_shinystan_demo}}
}

