% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/launch_shinystan.R
\name{launch_shinystan}
\alias{launch_shinystan}
\title{Launch shinyStan app}
\usage{
launch_shinystan(object)
}
\arguments{
\item{object}{An object of class \code{shinystan} or \code{stanfit}.
See \code{\link[shinyStan]{as.shinystan}} for how to easily convert other types
of objects to \code{shinystan} objects.}
}
\description{
Launch shinyStan app
}
\details{
If \code{object} is a \code{stanfit} object then, in addition to launching
the shinyStan app, this function creates an S4 object of class \code{shinystan} in
the Global Environment.
}
\note{
Unless you are using RStudio, \code{launch_shinystan} will open the app
in your system's default web browser. For RStudio users \strong{shinyStan} will
launch in RStudio's (pop-up) Viewer pane by default. If you prefer to use \strong{shinyStan}
in your web browser (or if you are having trouble with the RStudio Viewer pane) you
can click on 'Open in Browser' at the top of the Viewer pane.
}
\examples{
\dontrun{
# If X is a shinystan object or stanfit object then
# to launch the app just run
launch_shinystan(X)

# If X is not a shinystan or stanfit object then to launch the app first
# convert X to a shinystan object using the as.shinystan function
X_shinystan <- as.shinystan(X)
launch_shinystan(X_shinystan)
}
}
\seealso{
\code{\link[shinyStan]{as.shinystan}}, \code{\link[shinyStan]{launch_shinystan_demo}}
}

