% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/launch_shinystan.R
\name{launch_shinystan}
\alias{launch_shinystan}
\title{ShinyStan app}
\usage{
launch_shinystan(object, rstudio = getOption("shinystan.rstudio"), ...)
}
\arguments{
\item{object}{An object of class shinystan, stanfit, or
\code{stanreg}. See \code{\link{as.shinystan}} for converting other
objects to a shinystan object (sso).}

\item{rstudio}{For RStudio users, should the app launch in RStudio's Viewer?
The default is to launch the app in the user's default web browser instead
of the RStudio viewer unless the user has set
\code{options(shinystan.rstudio = TRUE)}, in which case the default is to
launch in RStudio's Viewer.}

\item{...}{Optional arguments to pass to \code{\link[shiny]{runApp}}.}
}
\value{
An S4 shinystan object.
}
\description{
ShinyStan app
}
\examples{
\dontrun{
# If X is a stanfit object (or shinystan object (sso))
X_sso <- launch_shinystan(X)

# If X is not an sso or stanfit object
X_sso <- launch_shinystan(as.shinystan(X, model_name = "Example"))
}
}
\seealso{
\code{\link{as.shinystan}},
  \code{\link{launch_shinystan_demo}}, \code{\link{shinystan_options}}
}

