% Generated by roxygen2 (4.0.2): do not edit by hand
\name{launch_SHINYstan_demo}
\alias{launch_SHINYstan_demo}
\title{Launch SHINYstan app in demo mode}
\usage{
launch_SHINYstan_demo(demo_name)
}
\arguments{
\item{demo_name}{Character string giving the name of the Stan demo model.
See \strong{Note} below for the available demos. If \code{demo_name} is
omitted it will default to the \code{eight_schools} Stan example model.}
}
\description{
This function will create an S4 object of class \code{SHINYstanfit} in the Global Environment
and then launch the SHINYstan app using a pre-loaded example model.
}
\note{
List of available demo models:
\describe{
  \item{\code{air}}{Berkson measurement error}
  \item{\code{beetles}}{Logit model}
  \item{\code{eight_schools}}{Multi-level linear model}
  \item{\code{pumps}}{Conjugate gamma-Poisson hierarchical model}
}
}
\examples{
\dontrun{
launch_SHINYstan_demo("eight_schools")
}
}
\seealso{
\code{\link[rstan]{stan_demo}}
}

