% Generated by roxygen2 (4.0.2): do not edit by hand
\name{launch_shinystan_demo}
\alias{launch_shinystan_demo}
\title{Launch SHINYstan app in demo mode}
\usage{
launch_shinystan_demo(demo_name = "eight_schools")
}
\arguments{
\item{demo_name}{Character string giving the name of the Stan demo model.
See \strong{Note} below for the available demos. If \code{demo_name} is
omitted it will default to the \code{eight_schools} Stan example model.}
}
\value{
In addition to launching the app, an object of class \code{shinystan}
is returned.
}
\description{
This function will create an S4 object of class \code{shinystan} in the Global Environment
and launch the SHINYstan app using a pre-loaded example model.
}
\note{
List of available demo models:
\describe{
  \item{\code{air}}{Berkson measurement error}
  \item{\code{beetles}}{Logit model}
  \item{\code{eight_schools}}{Multi-level linear model}
  \item{\code{pumps}}{Conjugate gamma-Poisson hierarchical model}
}
}
\examples{
\dontrun{
launch_shinystan_demo() # launches "eight_schools" demo by default
launch_shinystan_demo("air")
}
}

