% Generated by roxygen2 (4.0.2): do not edit by hand
\name{launch_SHINYstan_demo}
\alias{launch_SHINYstan_demo}
\title{Launch SHINYstan app in demo mode}
\usage{
launch_SHINYstan_demo()
}
\description{
This function will create an S4 object of class \code{SHINYstanfit} in the Global Environment
and then launch the SHINYstan app using the pre-loaded Stan example model "schools".
}
\seealso{
\code{\link[rstan]{stan_demo}}
}

