% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shinystan.R
\name{launch_shinystan_demo}
\alias{eight_schools}
\alias{launch_shinystan_demo}
\title{ShinyStan demo}
\usage{
launch_shinystan_demo(demo_name = "eight_schools",
  rstudio = getOption("shinystan.rstudio"), ...)
}
\arguments{
\item{demo_name}{The name of the demo. Currently \code{"eight_schools"} is 
the only option, but additional demos will be available in future releases.
\describe{
\item{\code{eight_schools}}{Hierarchical meta-analysis model. See
\emph{Meta Analysis} chapter of the Stan manual (chapter 11.2 in version 2.9),
\url{http://mc-stan.org/documentation/}.}
}}

\item{rstudio}{Only relevant for RStudio users. The default 
(\code{rstudio=FALSE}) is to launch the app in the default web browser 
rather than RStudio's pop-up Viewer. Users can change the default to 
\code{TRUE} by setting the global option \code{options(shinystan.rstudio = 
TRUE)}.}

\item{...}{Optional arguments to pass to \code{\link[shiny]{runApp}}.}
}
\value{
An S4 shinystan object.
}
\description{
ShinyStan demo
}
\examples{
\dontrun{
# launch demo but don't save a shinystan object
launch_shinystan_demo() 

# launch demo and save the shinystan object for the demo 
sso_demo <- launch_shinystan_demo()
}

}
\seealso{
\code{\link{launch_shinystan}} for launching ShinyStan using your 
  own shinystan objects
  
  \code{\link{as.shinystan}} to create shinystan objects
}

