% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/launch_shinystan_demo.R
\name{launch_shinystan_demo}
\alias{launch_shinystan_demo}
\title{Launch shinyStan app in demo mode}
\usage{
launch_shinystan_demo()
}
\value{
In addition to launching the \strong{shinyStan} app an S4 object of
  class \code{shinystan} is returned.
}
\description{
Launch shinyStan app in demo mode
}
\details{
Unless you are using RStudio, \code{launch_shinystan} will open your
  system's default web browser. For RStudio users \strong{shinyStan} will
  launch in RStudio's (pop-up) Viewer pane. If you prefer to use
  \strong{shinyStan} in your web browser (or if you are having trouble with
  the RStudio Viewer pane) you can click on 'Open in Browser' at the top of
  the Viewer pane.
}
\examples{
\dontrun{
launch_shinystan_demo()
}
}
\seealso{
\code{\link[shinyStan]{launch_shinystan}},
  \code{\link[shinyStan]{as.shinystan}},
}

