% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/launch_shinystan_demo.R
\name{launch_shinystan_demo}
\alias{launch_shinystan_demo}
\title{ShinyStan demo}
\usage{
launch_shinystan_demo(rstudio = getOption("shinystan.rstudio"), ...)
}
\arguments{
\item{rstudio}{Only relevant for RStudio users. If \code{TRUE} (the default),
the app will launch in RStudio's pop-up Viewer rather than the default web
browser. Users can change the default to \code{FALSE} by setting the global
option \code{options(shinystan.rstudio = FALSE)}.}

\item{...}{Optional arguments to pass to \code{\link[shiny]{runApp}}.}
}
\value{
An S4 shinystan object.
}
\description{
ShinyStan demo
}
\examples{
\dontrun{
# launch demo but don't save a shinystan object
launch_shinystan_demo()

# launch demo and save the shinystan object for the demo
ssdemo <- launch_shinystan_demo()
}
}
\seealso{
\code{\link{launch_shinystan}}, \code{\link{as.shinystan}}
}

