% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/launch_shinystan_demo.R
\name{launch_shinystan_demo}
\alias{launch_shinystan_demo}
\title{ShinyStan demo}
\usage{
launch_shinystan_demo(rstudio = FALSE, ...)
}
\arguments{
\item{rstudio}{For RStudio users, should the app launch in RStudio's Viewer?
The default (FALSE) is to launch the app in the default web browser.}

\item{...}{Optional arguments to pass to \code{\link[shiny]{runApp}}.}
}
\value{
An S4 shinystan object.
}
\description{
ShinyStan demo
}
\examples{
\dontrun{
launch_shinystan_demo()
ssdemo <- launch_shinystan_demo()
}
}
\seealso{
\code{\link{launch_shinystan}},
  \code{\link{as.shinystan}},
}

