% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shinystan.R
\name{launch_shinystan_demo}
\alias{launch_shinystan_demo}
\alias{eight_schools}
\alias{mt_cars_fullrank}
\alias{mt_cars_meanfield}
\title{'ShinyStan' demo}
\usage{
launch_shinystan_demo(
  demo_name = "eight_schools",
  rstudio = getOption("shinystan.rstudio", FALSE),
  quiet = getOption("shinystan.quiet", TRUE),
  ...
)
}
\arguments{
\item{demo_name}{The name of the demo. Currently \code{"eight_schools"} 
is the default. A small variaty of options are currently added to test 
the behavior under different algorithm specifications.
  \describe{
  \item{\code{eight_schools}}{Hierarchical meta-analysis model. See
   \emph{Meta Analysis} chapter of the 'Stan' manual
   \url{http://mc-stan.org/users/documentation/}.}
  \item{\code{mt_cars_fullrank}}{Linear regression using fullrank method. See
   the example in \code{stan_glm} from the \pkg{rstanarm}
  }
  \item{\code{mt_cars_meanfield}}{Linear regression using meanfield method. See
   the example in \code{stan_glm} from the \pkg{rstanarm} and the algorithm 
   documentation there for the specifics on meanfield.
  }
  }}

\item{rstudio}{Only relevant for 'RStudio' users. The default (\code{FALSE})
is to launch the app in the user's default web browser rather than the
pop-up Viewer provided by 'RStudio'. Users can change the default to
\code{TRUE} by setting the global option \code{options(shinystan.rstudio =
TRUE)}.}

\item{quiet}{Should output from Shiny to the R console be suppressed when
possible? The default is \code{TRUE}. Users can change the default to
\code{FALSE} by setting the global option \code{options(shinystan.quiet =
FALSE)}.}

\item{...}{Optional arguments passed to \code{\link[shiny]{runApp}}.}
}
\value{
An S4 shinystan object.
}
\description{
'ShinyStan' demo
}
\examples{
\dontrun{
# launch demo but don't save a shinystan object
launch_shinystan_demo()

# launch demo and save the shinystan object for the demo
sso_demo <- launch_shinystan_demo()
}

}
\seealso{
\code{\link{launch_shinystan}} to launch the 'ShinyStan' interface 
  using a particular \code{shinystan} object.

\code{\link{as.shinystan}} for creating \code{shinystan} objects.
}
