% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/launch_shinystan_demo.R
\name{launch_shinystan_demo}
\alias{launch_shinystan_demo}
\title{ShinyStan demo}
\usage{
launch_shinystan_demo(rstudio = getOption("shinystan.rstudio"), ...)
}
\arguments{
\item{rstudio}{Only relevant for RStudio users. The default
(\code{rstudio=FALSE}) is to launch the app in the default web browser
rather than RStudio's pop-up Viewer. Users can change the default to
\code{TRUE} by setting the global option \code{options(shinystan.rstudio =
TRUE)}.}

\item{...}{Optional arguments to pass to \code{\link[shiny]{runApp}}.}
}
\value{
An S4 shinystan object.
}
\description{
ShinyStan demo
}
\examples{
\dontrun{
# launch demo but don't save a shinystan object
launch_shinystan_demo()

# launch demo and save the shinystan object for the demo
ssdemo <- launch_shinystan_demo()
}
}
\seealso{
\code{\link{launch_shinystan}}, \code{\link{as.shinystan}}
}

