% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shinystan.R
\name{launch_shinystan_demo}
\alias{launch_shinystan_demo}
\alias{eight_schools}
\title{'ShinyStan' demo}
\usage{
launch_shinystan_demo(demo_name = "eight_schools",
  rstudio = getOption("shinystan.rstudio"), quiet = TRUE, ...)
}
\arguments{
\item{demo_name}{The name of the demo. Currently \code{"eight_schools"} is 
the only option, but additional demos may be available in future releases.
\describe{
\item{\code{eight_schools}}{Hierarchical meta-analysis model. See 
 \emph{Meta Analysis} chapter of the 'Stan' manual 
 \url{http://mc-stan.org/users/documentation/}.}
}}

\item{rstudio}{Only relevant for 'RStudio' users. The default (\code{FALSE})
is to launch the app in the user's default web browser rather than the 
pop-up Viewer provided by 'RStudio'. Users can change the default to
\code{TRUE} by setting the global option \code{options(shinystan.rstudio =
TRUE)}.}

\item{quiet}{Should output from Shiny to the R console be suppressed? Default is \code{TRUE}.}

\item{...}{Optional arguments passed to \code{\link[shiny]{runApp}}.}
}
\value{
An S4 shinystan object.
}
\description{
'ShinyStan' demo
}
\examples{
\dontrun{
# launch demo but don't save a shinystan object
launch_shinystan_demo() 

# launch demo and save the shinystan object for the demo 
sso_demo <- launch_shinystan_demo()
}

}
\seealso{
\code{\link{launch_shinystan}} to launch the 'ShinyStan' interface 
  using a particular \code{shinystan} object.

\code{\link{as.shinystan}} for creating \code{shinystan} objects.
}
