% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mcmc2shinystan}
\alias{mcmc2shinystan}
\title{Convert an object of class \code{mcmc.list} to class \code{shinystan}}
\usage{
mcmc2shinystan(X, model_name = deparse(substitute(X)), burnin = 0,
  param_dims = list())
}
\arguments{
\item{X}{An \code{mcmc.list}}

\item{model_name}{A character string given a name for the model.}

\item{param_dims}{The dimensions for all parameters. A named list.}

\item{The}{number of warmup/burnin iterations. Not needed if the samples don't include any of the burnin.}
}
\description{
Convert an object of class \code{mcmc.list} to class \code{shinystan}
}
\seealso{
\code{\link[coda]{mcmc.list}}
}

