% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_code.R
\name{model_code}
\alias{model_code}
\title{Add model code to a shinystan object}
\usage{
model_code(sso, code)
}
\arguments{
\item{sso}{A shinystan object.}

\item{code}{The code you want to add. See \strong{Details} below for
formatting instructions.}
}
\value{
The shinystan object \code{sso} with \code{code} in the
  \code{model_code} slot.
}
\description{
Add model code to a shinystan object
}
\details{
\code{code} should be a character string that can be used as an
  argument to \code{cat}. See \strong{Examples}, below.
}
\note{
This is intended for users who did not run their models using
  \pkg{rstan}. For \pkg{rstan} users the model code will be automatically
  available.
}
\examples{
\dontrun{
# Some JAGS-style code we might want to add
my_code <- "
 model {
   for (i in 1:length(Y)) {
     Y[i] ~ dpois(lambda[i])
     log(lambda[i]) <- inprod(X[i,], theta[])
   }
   for (j in 1:J) {
     theta[j] ~ dt(0.0, 1.0, 1.0)
   }
 }
"

# Add the code to a shinystan object sso
sso <- model_code(sso, my_code)

}
}
\seealso{
\code{cat}
}

