% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/notes.R
\name{notes}
\alias{notes}
\title{Additional or replacement notes}
\usage{
notes(sso, notes, replace = FALSE)
}
\arguments{
\item{sso}{shinystan object.}

\item{notes}{A character vector of additional or replacement notes.}

\item{replace}{If \code{TRUE} the existing notes are overwritten by
\code{notes}. Otherwise the new notes are appended to the existing notes.}
}
\value{
\code{sso}, updated. See Examples.
}
\description{
Notes are viewable on ShinyStan's Notepad page
}
\examples{
\dontrun{
sso <- notes(sso, "new note")
sso <- notes(sso, c("new note", "another new note"), replace = TRUE)
}
}
\seealso{
\code{\link{as.shinystan}}
}

