% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sso-helpers.R
\name{rename_model}
\alias{rename_model}
\title{Change the model name associated with a shinystan object}
\usage{
rename_model(sso, new_model_name)
}
\arguments{
\item{sso}{The shinystan object to use.}

\item{new_model_name}{Character string giving the new model name to use.}
}
\value{
sso, updated.
}
\description{
Change the model name associated with a shinystan object
}
\examples{
# Using example shinystan object 'eight_schools'
sso <- eight_schools
sso@model_name
sso <- rename_model(sso, "New name")
sso@model_name

}
\seealso{
\code{\link{as.shinystan}}
}

