% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rename_model.R
\name{rename_model}
\alias{rename_model}
\title{Change the model name associated with a \code{shinystan} object}
\usage{
rename_model(sso, new_model_name)
}
\arguments{
\item{sso}{The \code{shinystan} object (\code{sso}) you want to edit.}

\item{new_model_name}{Character string giving the new model name to use.}
}
\description{
Change the model name associated with a \code{shinystan} object
}
\examples{
\dontrun{
# Assume X is a shinystan object
X <- rename_model(X, "Any character string")
}
}
\seealso{
\code{\link{as.shinystan}}
}

