% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_model.R
\name{rename_model}
\alias{rename_model}
\title{Change the model name associated with a shinystan object}
\usage{
rename_model(sso, new_model_name)
}
\arguments{
\item{sso}{shinystan object.}

\item{new_model_name}{Character string giving the new model name to use.}
}
\value{
sso, updated.
}
\description{
Change the model name associated with a shinystan object
}
\examples{
\dontrun{
sso <- rename_model(sso, "new name for model")
}

}
\seealso{
\code{\link{as.shinystan}}
}

