% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinystan-objects.R
\docType{class}
\name{shinystan-class}
\alias{shinystan-class}
\alias{shinystan}
\title{S4 \code{shinystan} objects}
\description{
See \code{\link{as.shinystan}} for documentation on creating
  \code{shinystan} objects and \code{\link{eight_schools}} for an example
  object.
}
\section{Slots}{

\describe{
\item{\code{model_name}}{(\code{"character"}) Model name.}

\item{\code{param_names}}{(\code{"character"}) Parameter names.}

\item{\code{param_dims}}{(\code{"list"}) Parameter dimensions.}

\item{\code{posterior_sample}}{(\code{"array"}) MCMC sample.}

\item{\code{summary}}{(\code{"matrix"}) Summary stats for \code{posterior_sample}.}

\item{\code{monitor_summary}}{(\code{"data.frame"}) Summary stats obtained using
\code{monitor} function from \pkg{rstan}.}

\item{\code{sampler_params}}{(\code{"list"}) Sampler parameters (for certain Stan
models only).}

\item{\code{n_chain}}{(\code{"integer"}) Number of chains.}

\item{\code{n_iter}}{(\code{"integer"}) Number of iterations per chain.}

\item{\code{n_warmup}}{(\code{"integer"}) Number of warmup iterations per chain.}

\item{\code{user_model_info}}{(\code{"character"}) Notes to display on the
\strong{Notepad} page in the 'ShinyStan' GUI.}

\item{\code{model_code}}{(\code{"character"}) Model code to display on the
\strong{Model Code} page in the 'ShinyStan' GUI.}

\item{\code{stan_used}}{(\code{"logical"}) Logical indicator if stan was used.}

\item{\code{stan_method}}{(\code{"character"}) If stan was used, which method.}

\item{\code{stan_algorithm}}{(\code{"character"}) If stan was used, which algorithm.}

\item{\code{sso_version}}{(\code{"character"}) What version of shinystan was used
to create the shinystan object.}

\item{\code{misc}}{(\code{"list"}) Miscellaneous, for internal use. Will be
removed in future releases.}
}}

\references{
Muth, C., Oravecz, Z., and Gabry, J. (2018)
User-friendly Bayesian regression modeling: A tutorial with rstanarm and shinystan.
\emph{The Quantitative Methods for Psychology}. 14(2), 99--119.
\url{https://www.tqmp.org/RegularArticles/vol14-2/p099/p099.pdf}
}
\seealso{
\code{\link{as.shinystan}} for creating \code{shinystan} objects.

\code{\link{drop_parameters}} to remove parameters from a
  \code{shinystan} object.

\code{\link{generate_quantity}} to add a new quantity to a
  \code{shinystan} object.

\code{\link{shinystan-metadata}} to view or change metadata
  associated with a \code{shinystan} object.
}
