% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sso-helpers.R
\name{sso_info}
\alias{sso_info}
\title{Print summary of shinystan object}
\usage{
sso_info(sso)
}
\arguments{
\item{sso}{The shinystan object to use.}
}
\description{
Prints basic summary info including number of parameters, chains, iterations, 
warmup iterations, etc.
}
\examples{
# Using example shinystan object 'eight_schools'
sso_info(eight_schools)

}

