% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sso_info.R
\name{sso_info}
\alias{sso_info}
\title{Print summary of shinystan object}
\usage{
sso_info(sso)
}
\arguments{
\item{sso}{A \code{shinystan} object.}
}
\description{
Prints basic summary info including number of parameters, chains, iterations,
warmup iterations, etc.
}

