% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sso.R
\name{update_sso}
\alias{update_sso}
\title{Update an object created by an earlier version of shinystan}
\usage{
update_sso(old_sso)
}
\arguments{
\item{old_sso}{An old shinystan object to update.}
}
\value{
sso, updated.
}
\description{
Before you can use an old shinystan object (sso) with the new version of 
shinystan you might need to run \code{update_sso}. The updated sso
will then have all the slots that will be accessed by the ShinyStan app.
}
\examples{
\dontrun{
sso_new <- update_sso(sso_old)
}  

}

